+++ Config/BuildSettings.swift
@@ -201,5 +201,7 @@ final class BuildSettings: NSObject {
     
     // MARK: - Authentication Screen
     static let authScreenShowRegister = true
+    static let authScreenShowPhoneNumber = true
+    static let authScreenShowForgotPassword = true
     static let authScreenShowCustomServerOptions = true
 }