+++ CHANGES.rst
@@ -7,6 +7,7 @@ Features:
 Improvements:
  * Config fixes.
  * Introduce TableViewSections. Refactor RoomSettingsViewController & SettingsViewController.
+ * AuthenticationVC: Make forgot password button and phone number text field configurable.
 
 Bugfix:
  * Fix biometry name null case (#3551).